projects/prescription_create/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf_yasg
pika